requirements.txt
fastapi>=0.124.1,<1.0.0
uvicorn[standard]>=0.34.0,<1.0.0
pydantic>=2.12.0,<3.0.0
pydantic-settings>=2.5.2
httpx==0.27.2
slowapi==0.1.9
python-dotenv==1.0.1
structlog==24.4.0
google-adk>=1.25.0
pyyaml>=6.0
opentelemetry-api>=1.30
opentelemetry-sdk>=1.30
opentelemetry-exporter-gcp-trace>=1.8.0
opentelemetry-instrumentation-fastapi>=0.48b0
